--- cert_viewer/errors.py
@@ -1,7 +1,0 @@
-class Error(Exception):
-    """Base class for exceptions in this module"""
-    pass
-
-
-class InvalidTransactionError(Error):
-    pass
