--- cert_viewer/models.py
@@ -1,4 +1,0 @@
-from collections import namedtuple
-
-UserData = namedtuple('UserData', ['pubkey', 'email', 'degree', 'comments', 'first_name', 'last_name',
-                                   'street_address', 'city', 'state', 'zip_code', 'country'])
